requirements.txt
transformers==4.26.0
Pillow==9.2.0
requests==2.28.1
